Fix compilation warnings

+++ ja2lib/Strategic/TownMilitia.c
@@ -1131,7 +1131,7 @@ void BuildMilitiaPromotionsString(CHAR16 *str, size_t bufSize) {
 bool HasNewMilitiaPromotions() { return _st.gbMilitiaPromotions > 0; }
 
 BOOLEAN DoesPlayerHaveAnyMilitia() {
-  INT16 sX, sY;
+  u8 sX, sY;
 
   // run through list of towns that might have militia..if any return TRUE..else return FALSE
   for (sX = 1; sX < MAP_WORLD_X - 1; sX++) {
